Changed PV depth to uint8 to be consistent

+++ src/details/pv_table.hpp
@@ -14,19 +14,19 @@ namespace details
 
 struct pv_table
 {
-    static constexpr std::size_t MAX_DEPTH { 64 };
+    static constexpr std::uint8_t MAX_DEPTH { 64 };
 
     // Only clears the length storage.
     void clear() noexcept;
 
     // Called when we find a new best move at a given ply (distance from root).
-    void update(std::size_t ply, std::uint32_t move);
+    void update(std::uint8_t ply, std::uint32_t move);
 
     // Retrieve root PV after a search.
     std::span<const std::uint32_t> get_pv() const noexcept;
 
     std::array<std::array<std::uint32_t, MAX_DEPTH>, MAX_DEPTH> table;
-    std::array<std::size_t, MAX_DEPTH> lengths {};
+    std::array<std::uint8_t, MAX_DEPTH> lengths {};
 };
 
 // ####################################
@@ -38,7 +38,7 @@ inline void pv_table::clear() noexcept
     lengths.fill(0);
 }
 
-inline void pv_table::update(std::size_t ply, std::uint32_t move)
+inline void pv_table::update(std::uint8_t ply, std::uint32_t move)
 {
     table[ply][0] = move;
